==> 2b_scrape_youtube_comments.ipynb <==
Continues from 2_youtube_search.ipynb

from googleapiclient.discovery import build
from IPython.display import YouTubeVideo
from pathlib import Path
import pandas as pd
import json
from tqdm.auto import tqdm
import re

### Config

PRODUCT = "wireless over-ear headphones"
MAX_COMMENTS = 100

# Ensure key has Youtube Data API enabled in Google Cloud Console.
GOOGLE_YOUTUBE_KEY = "AIzaSyCF-hCRmYwVA4v6CD5bAWWvG-m8rwQohmY"

DATA_DIR = Path("session") / PRODUCT
META_PATH = DATA_DIR / "stage_1.json"
assert META_PATH.exists(), "Run 1_describe_product.ipynb first!"

YOUTUBE_DIR = DATA_DIR / "youtube"
VID_META_PATH = YOUTUBE_DIR / "vid_meta.json"
assert VID_META_PATH.exists(), "Run 2_youtube_search.ipynb first!"

OUT_DIR = YOUTUBE_DIR / "raw_comments"
OUT_DIR.mkdir(exist_ok=True, parents=True)

### Operations

#### Setup

youtube = build("youtube", "v3", developerKey=GOOGLE_YOUTUBE_KEY)

with open(VID_META_PATH, "r") as f:
    vid_meta = json.load(f)

display(vid_meta)

#### Retrieve Comments on Each Competitor

competitor_filemap = {}

for competitor, viddata in vid_meta.items():
    print(f"Retrieving comments for {competitor}.")
    name = re.sub("[^0-9a-zA-Z]+", "_", competitor)
    competitor_csv = OUT_DIR / f"{name}.csv"
    competitor_filemap[competitor] = f"{name}.csv"

    videos = []
    for category, vids in viddata.items():
        videos.extend([(category, vid) for vid in vids])

    data = []
    for cat, vid in tqdm(videos):
        req = youtube.commentThreads().list(
            part="snippet,replies",
            videoId=vid,
            maxResults=100,
            textFormat="plainText",
        )
        
        this_data = []
        while req and len(this_data) < MAX_COMMENTS:
            res = req.execute()
            req = youtube.commentThreads().list_next(req, res)

            for thread in res["items"]:
                comments = [thread["snippet"]["topLevelComment"]]
                if "replies" in thread:
                    comments.extend(thread["replies"]["comments"])
                # print("Number of comments in thread:", len(comments))
            
                for com in comments:
                    this_data.append(
                        {
                            "product": competitor,
                            "category": cat,
                            "id": com["id"],
                            "videoId": vid,
                            "parentId": thread["id"],
                            "isReply": thread["id"] != com["id"],
                            "authorName": com["snippet"]["authorDisplayName"],
                            "authorId": com["snippet"]["authorChannelId"]["value"],
                            "text": com["snippet"]["textDisplay"],
                            "likes": com["snippet"]["likeCount"],
                            "updatedAt": com["snippet"]["updatedAt"],
                        }
                    )

        data.extend(this_data)

    df = pd.DataFrame(data)
    df.to_csv(competitor_csv, index=False)
    print(f"Saved {len(df)} comments to {competitor_csv}.")

with open(YOUTUBE_DIR / "comment_filemap.json", "w") as f:
    json.dump(competitor_filemap, f, indent=2)
