==> 3_gemini_video_analysis.ipynb <==
Continues from 2_youtube_search.ipynb

from google import genai
from google.genai import types
from IPython.display import YouTubeVideo, Markdown
from pathlib import Path
from typing import *
from pydantic import BaseModel, Field, create_model
from devtools import debug
import json
from tqdm.auto import tqdm

### Config
Remember to change PRODUCT to match the current analysis session!

PRODUCT = "wireless over-ear headphones"
TOPIC = "review"  # Need to match TOPICS from 2_youtube_search.ipynb!

# Should be from Google AI Studio.
GOOGLE_AI_KEY = "AIzaSyDAlPx7St5BUXqlwiqFKvlT-Sc2dnTT4Jc"
# 2.5 Flash is good tradeoff between better vid understanding and free 500 RPD.
GOOGLE_AI_MODEL = "gemini-2.5-flash-preview-04-17"
# GOOGLE_AI_MODEL = "gemini-2.5-pro-exp-03-25"

DATA_DIR = Path("session") / PRODUCT
META_PATH = DATA_DIR / "stage_1.json"
assert META_PATH.exists(), "Run 1_describe_product.ipynb first!"

YOUTUBE_DIR = DATA_DIR / "youtube"
VID_META_PATH = YOUTUBE_DIR / "vid_meta.json"
assert VID_META_PATH.exists(), "Run 2_youtube_search.ipynb first!"

TRANSCRIPT_PATH = YOUTUBE_DIR / "transcripts.json"
ANALYSIS_PATH = YOUTUBE_DIR / "analysis.json"

### Operations

#### Setup

client = genai.Client(api_key=GOOGLE_AI_KEY)

with open(VID_META_PATH, "r") as f:
    metadata = json.load(f)

with open(META_PATH, "r") as f:
    stage1_meta = json.load(f)

videos = {}
for competitor in metadata:
    videos[competitor] = metadata[competitor][TOPIC]

display(videos)

#### Generate Full Transcript
Getting the full transcript first is useful for manual analysis (if needed). More
importantly, it ensures the model doesn't miss any details and minimizes hallucination.

prompt_transcript = f"""\
Provide the full section by section transcript of this video. Include all possible
details such as:
- who said what to whom
- on-screen text and charts
- music and sound effects
- change in object states

Give the transcript in neatly formatted markdown sections with headers.\
"""

display(Markdown(prompt_transcript))

def transcribe_video(vid):
    contents = [
        types.Content(
            role="user",
            parts=[
                types.Part.from_uri(
                    file_uri=f"https://youtu.be/{vid}", mime_type="video/*"
                ),
                types.Part.from_text(text=prompt_transcript),
            ],
        )
    ]

    resp = client.models.generate_content(
        model=GOOGLE_AI_MODEL,
        contents=contents,
    )

    return resp.text

if TRANSCRIPT_PATH.exists():
    with open(TRANSCRIPT_PATH, "r") as f:
        transcripts_meta = json.load(f)

    print("Previous transcript found:")
    print("Topic:", transcripts_meta["vid_topic"])
    print("Model:", transcripts_meta["model"])

    transcripts = transcripts_meta["products"]
else:
    transcripts = {}
    transcripts_meta = {
        "vid_topic": TOPIC,
        "model": GOOGLE_AI_MODEL,
        "products": transcripts,
    }
    print("No previous transcripts found.")

# Will take approx one minute per new video!
for competitor, vids in videos.items():
    for vid in tqdm(vids, desc=competitor):
        transcripts[competitor] = transcripts.get(competitor, {})
        text = transcripts[competitor].get(vid, None)
        if text is not None:
            print(f"Already transcribed {vid}")
            continue

        while text is None:
            print(f"Transcribing {vid}...")
            text = transcribe_video(vid)
            if text is None:
                print("Transcription failed, retrying...")
                continue
            print(f"Transcribed {vid}")
            transcripts[competitor][vid] = text

transcripts_meta["products"] = transcripts

with open(TRANSCRIPT_PATH, "w") as f:
    json.dump(transcripts_meta, f, indent=2)

#### Preview Transcript

display(Markdown(next(iter(next(iter(transcripts.values())).values()))))

#### Analyze Strengths and Weaknesses
Each strength and weakness is based around a specific design metric. We also leave
an others section for any additional notes.

metrics = stage1_meta["metrics"] + ["others"]


class Feedback(BaseModel):
    metric: Literal[*metrics] = Field(description="Metric being evaluated")
    rationale: str = Field(description="Reasoning for the given evaluation")
    rating: int = Field(ge=-3, le=3, description="Rating from -3 to 3")
    component: str = Field(
        description="Specific component responsible for the evaluation"
    )
    evidence: List[str] = Field(
        description="List of evidence to support the evaluation"
    )


Evaluation = create_model(
    "Evaluation",
    **{
        m: Annotated[List[Feedback], Field(description=f"Evaluation of {m}")]
        for m in metrics
    },
)


class ProductEvalResult(BaseModel):
    product: str = Field(description="Product name")
    evaluation: Evaluation = Field(description="Evaluation of the product")

def analyze_video(competitor, vid, transcript):
    prompt = f"""\
Refer only to the video and its transcript above. Give feedback on all strengths \
and weaknesses of {competitor} {PRODUCT} with respect to the following design metrics:
{"\n".join(f"- {s}" for s in stage1_meta["metrics"])}

Be concise but also as specific as possible. Write down which exact component of \
{competitor} is responsible for the feedback, and quantitative or qualitative values \
for why it is good or bad. For example:
    
```json
{{
    "metric": "durability",
    "rationale": "...",
    "rating": 3,
    "component": "carbon fiber reinforced hinges",
    "evidence": [
        "1,000,000 open/close cycles"
    ]
}}
```

When answering, use the following JSON schema:

{ProductEvalResult.model_json_schema()}
"""

    contents = [
        types.Content(
            role="user",
            parts=[
                types.Part.from_uri(
                    file_uri=f"https://youtu.be/{vid}", mime_type="video/*"
                ),
                types.Part.from_text(text=prompt_transcript),
            ],
        ),
        types.Content(
            role="model",
            parts=[
                types.Part.from_text(text=transcript),
            ],
        ),
        types.Content(
            role="user",
            parts=[
                types.Part.from_text(text=prompt),
            ],
        ),
    ]
    resp = client.models.generate_content(
        model=GOOGLE_AI_MODEL,
        contents=contents,
        config=types.GenerateContentConfig(
            responseMimeType="application/json",
            responseSchema=ProductEvalResult,
        ),
    )
    return resp.parsed

if ANALYSIS_PATH.exists():
    with open(ANALYSIS_PATH, "r") as f:
        analysis_meta = json.load(f)

    print("Previous analysis found:")
    print("Topic:", analysis_meta["vid_topic"])
    print("Model:", analysis_meta["model"])

    analysis = analysis_meta["products"]
else:
    analysis = {}
    analysis_meta = {"vid_topic": TOPIC, "model": GOOGLE_AI_MODEL, "products": analysis}
    print("No previous analysis found.")

# Will take approx one minute per new video!
for competitor in transcripts:
    for vid, transcript in tqdm(transcripts[competitor].items(), desc=competitor):
        analysis[competitor] = analysis.get(competitor, {})
        result = analysis[competitor].get(vid, None)
        if result is not None:
            print(f"Already analyzed {vid}")
            continue

        while result is None:
            print(f"Analyzing {vid}...")
            result = analyze_video(competitor, vid, transcript)
            if result is None:
                print("Analysis failed, retrying...")
                continue
            print(f"Analyzed {vid}")
            analysis[competitor][vid] = result.model_dump()

analysis_meta["products"] = analysis


with open(ANALYSIS_PATH, "w") as f:
    json.dump(analysis_meta, f, indent=2)

#### Preview Analysis

_ = debug(next(iter(next(iter(analysis.values())).values())))
