==> 3b_summarize_comments.ipynb <==
Continues from 3_gemini_video_analysis.ipynb.

from google import genai
from google.genai import types
from IPython.display import YouTubeVideo, Markdown
from pathlib import Path
from typing import *
from pydantic import BaseModel, Field, create_model
from devtools import debug
import json
from tqdm.auto import tqdm
import pandas as pd
import time

### Config

PRODUCT = "wireless over-ear headphones"

# Should be from Google AI Studio.
GOOGLE_AI_KEY = "AIzaSyDAlPx7St5BUXqlwiqFKvlT-Sc2dnTT4Jc"
# 2.0 Flash since free 1500 RPD and Gemma's structured output is disabled.
GOOGLE_AI_MODEL = "gemini-2.0-flash"

DATA_DIR = Path("session") / PRODUCT
META_PATH = DATA_DIR / "stage_1.json"
assert META_PATH.exists(), "Run 1_describe_product.ipynb first!"

REDDIT_FILEMAP_PATH = DATA_DIR / "reddit" / "competitor_map.json"
YOUTUBE_FILEMAP_PATH = DATA_DIR / "youtube" / "comment_filemap.json"
assert REDDIT_FILEMAP_PATH.exists(), "Run 2_scrape_reddit_reviews.ipynb first!"
assert YOUTUBE_FILEMAP_PATH.exists(), "Run 2_scrape_youtube_comments.ipynb first!"

OUT_DIR = DATA_DIR / "comment_analysis"
OUT_DIR.mkdir(exist_ok=True, parents=True)

### Operations

#### Setup

client = genai.Client(api_key=GOOGLE_AI_KEY)

with open(REDDIT_FILEMAP_PATH, "r") as f:
    reddit_filemap = json.load(f)

with open(YOUTUBE_FILEMAP_PATH, "r") as f:
    youtube_filemap = json.load(f)

with open(META_PATH, "r") as f:
    stage1_meta = json.load(f)

#### Load Reddit Comments

r_comments = {}

for competitor in reddit_filemap:
    filename = reddit_filemap[competitor]
    csv_path = DATA_DIR / "reddit" / "processed_comments" / f"{filename}.csv"
    try:
        df = pd.read_csv(csv_path)
    except Exception as e:
        print(f"Error reading {csv_path}: {e}")
        continue

    r_comments[competitor] = {}
    for metric, gdf in df.groupby("features"):
        r_comments[competitor][metric] = gdf["comment_body"].tolist()
        # print(competitor, metric, len(gdf))

# display(r_comments["Focal Bathys"]["Sound Quality"])

#### Load YouTube Comments

y_comments = {}

for competitor in youtube_filemap:
    filename = youtube_filemap[competitor]
    csv_path = DATA_DIR / "youtube" / "processed_comments" / filename
    try:
        df = pd.read_csv(csv_path)
    except Exception as e:
        print(f"Error reading {csv_path}: {e}")
        continue

    y_comments[competitor] = {}
    for metric, gdf in df.groupby("features"):
        y_comments[competitor][metric] = gdf["text"].tolist()
        # print(competitor, metric, len(gdf))

# display(y_comments["Focal Bathys"]["Sound Quality"])

#### Extract Feedback From Comments

EVIDENCE_LIMIT = 3
FEEDBACK_LIMIT = 3


def analyze_comments(competitor, metric, comments):
    class Feedback(BaseModel):
        rationale: str = Field(description="Reasoning for the given evaluation")
        rating: int = Field(ge=-3, le=3, description="Rating from -3 to 3")
        component: str = Field(
            description="Specific component responsible for the evaluation"
        )
        evidence: List[str] = Field(
            max_length=EVIDENCE_LIMIT,
            description="List of evidence to support the evaluation",
        )

    class EvalResult(BaseModel):
        product: str = Field(description="Product being evaluated")
        metric: str = Field(description="Metric being evaluated")
        feedbacks: List[Feedback] = Field(
            max_length=FEEDBACK_LIMIT, description=f"Feedback on {metric}"
        )

    prompt = f"""\
## Job Description
You are a product analyst. Your job is to analyze the comments given to the \
{competitor} {PRODUCT} on the basis of {metric} from Reddit and YouTube and extract useful feedback.

When writing feedback on {metric}, extract as much specific information as possible. \
Write down which exact component of \
{competitor} is responsible for the feedback, and quantitative or qualitative values \
for why it is good or bad. For example:
    
```json
{{
    "rationale": "...",
    "rating": 3,
    "component": "carbon fiber reinforced hinges",
    "evidence": [
        "1,000,000 open/close cycles"
    ]
}}
```

## Comments

The comments to analyze are given below:

{"\n\n".join(f"### Comment {i + 1}\n{comment}" for i, comment in enumerate(comments))}

## Response Schema

When giving feedback on the {metric} of {competitor} {PRODUCT}, use the following JSON schema:

{EvalResult.model_json_schema()}\
"""

    resp = client.models.generate_content(
        model=GOOGLE_AI_MODEL,
        contents=prompt,
        config=types.GenerateContentConfig(
            responseMimeType="application/json",
            responseSchema=EvalResult,
        ),
    )
    return resp.parsed


# Combine all comments from Reddit and YouTube
competitors = set(r_comments.keys()).union(set(y_comments.keys()))

all_comments = {}
for competitor in competitors:
    all_comments[competitor] = {}
    for metric in stage1_meta["metrics"]:
        all_comments[competitor][metric] = r_comments.get(competitor, {}).get(
            metric, []
        ) + y_comments.get(competitor, {}).get(metric, [])

tasks = []
for competitor, cdata in all_comments.items():
    for metric, comments in cdata.items():
        if len(comments) == 0:
            continue
        tasks.append((competitor, metric, comments))

for competitor, metric, comments in tqdm(tasks):
    result = None
    while result is None:
        try:
            result = analyze_comments(competitor, metric, comments)
            path = OUT_DIR / competitor / f"{metric}.json"
            path.parent.mkdir(exist_ok=True, parents=True)
            with open(path, "w") as f:
                json.dump(result.model_dump(), f, indent=2)

        except Exception as e:
            print(f"Error analyzing {competitor} {metric}: {e}")
            print("Retrying...")
            # We probably hit the per minute rate limit.
            time.sleep(10)
